Add a websocket ping task and automatic Discord reconnection
After the connection has been active for some time, I've seen the
websocket go into a fail state. The discord package code seems to try to
resume or restart the websocket in some way, but this failed to happen
properly, causing the bot to hang.  This is an issue that seems similar
to one that happened with beem, where, after a while, websockets entered
into a failed state yet never triggered an exception that could be
caught through the websockets code. One difference with the Discord
websocket issue is that an exception is being raised in the discord
package, yet its attempt at resuming doesn't seem successful.

In beem I created a ping task that calls the `.ping()` method
periodically to send data over the websocket. Doing so seemed to prevent
the websocket from failing in this way. I'm adopting the same approach
for now in Cerebot, doing a `.ping()` every 10 seconds and triggering a
full shutdown of the discord manager when an exception is raised in this
task. This shutdown will have the control loop delete the manager and
reinstantiate it to fully recreate the Discord connection. This is a bit
awkward, and the discord client should probably be moved to its own
property instead of having the manager inherit from the discord.Client
class itself. Furthermore some more investigation could be done to see
if some proper exception handling through the client API can lead to
better handling of these failures without maintaining this ping task at
all.