Properly handle errors in the websocket ping task
Three major errors: our exception logging message was broken, we called
the wrong method to disconnect the websocket, and we tried to yield to
the disconnect coroutine, which itself cancels our task! Fix these so
that websocket error from the ping task will correctly lead to
recreation of the discord client.